Write a test for signup route
