+++ backend/package.json
@@ -21,8 +21,10 @@
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
     "@types/node": "^22.10.1",
+    "@types/supertest": "^6.0.2",
     "dotenv": "^16.4.5",
     "nodemon": "^3.1.7",
+    "supertest": "^7.0.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.7.2"
   }
